+++ dantro/__init__.py
@@ -9,7 +9,7 @@ logging.basicConfig(format="%(levelname)-7s %(module)-12s %(message)s",
 log = logging.getLogger(__name__)
 
 # Define version variable
-__version__ = '0.4.0-pre.1'
+__version__ = '0.4.0-pre.2'
 # NOTE This should always be the same as in setup.py
 
 # TODO consider making some classes available here