+++ dantro/base.py
@@ -10,6 +10,7 @@ NOTE: These classes are not meant to be instantiated.
 """
 
 import abc
+import copy
 import logging
 import warnings
 import inspect
@@ -56,12 +57,9 @@ class AttrsMixin:
     def attrs(self, new_attrs):
         """Setter method for the container `attrs` attribute."""
         # Decide which class to use for attributes
-        if self._ATTRS_CLS is not None:
-            # Use the pre-defined one
-            AttrsCls = self._ATTRS_CLS
-        else:
-            # Use a default
-            AttrsCls = BaseDataAttrs
+        AttrsCls = (BaseDataAttrs
+                    if self._ATTRS_CLS is None
+                    else self._ATTRS_CLS)
 
         # Perform the initialisation
         log.debug("Using %s for attributes of %s",
@@ -86,19 +84,22 @@ class PathMixin:
     def parent(self, cont):
         """Associate a parent object with this container."""
         if self.parent is not None and cont is not None:
-            log.warning("A parent was already associated with %s '%s'! Will "
-                        "ignore this assignment.", self.classname, self.name)
-        else:
-            log.debug("Setting %s as parent of %s ...",
-                      cont.logstr if cont else None, self.logstr)
-            self._parent = cont
+            raise ValueError("A parent was already associated with {cls:} "
+                             "'{}'! Instead of manually setting the parent, "
+                             "use the functions supplied to manipulate "
+                             "members of this {cls:}."
+                             "".format(self.name, cls=self.classname))
+        
+        log.debug("Setting %s as parent of %s ...",
+                  cont.logstr if cont else None, self.logstr)
+        self._parent = cont
 
     @property
     def path(self) -> str:
         """Return the path to get to this container"""
         if self.parent is None:
             # At the top or no parent associated -> no reasonable path to give
-            return PATH_JOIN_CHAR
+            return self.name
         # else: not at the top, also need the parent's path
         return self.parent.path + PATH_JOIN_CHAR + self.name
 
@@ -128,20 +129,32 @@ class CollectionMixin:
 
 class ItemAccessMixin:
     """This Mixin class implements the methods needed for getting, setting,
-    and deleting items. It relays all calls forward to the data attribute.
+    and deleting items. It relays all calls forward to the data attribute, but
+    if given a list (passed down from above), it extracts it
     """
 
     def __getitem__(self, key):
         """Returns an item."""
+        key = self.__item_convert_key(key)
         return self.data[key]
 
     def __setitem__(self, key, val):
         """Sets an item."""
+        key = self.__item_convert_key(key)
         self.data[key] = val
 
     def __delitem__(self, key):
         """Deletes an item"""
+        key = self.__item_convert_key(key)
         del self.data[key]
+
+    def __item_convert_key(self, key):
+        """If given something that is not a list, just return that key"""
+        if isinstance(key, list):
+            if len(key) > 1:
+                return tuple(key)
+            return key[0]
+        return key
 
 
 class MappingAccessMixin(ItemAccessMixin, CollectionMixin):
@@ -292,7 +305,7 @@ class BaseDataAttrs(MappingAccessMixin, dantro.abc.AbstractDataAttrs):
 
     def _format_info(self) -> str:
         """A __format__ helper function: returns info about these attributes"""
-        return "{} attributes".format(len(self))
+        return "{} attribute(s)".format(len(self))
 
 
 # -----------------------------------------------------------------------------
@@ -350,18 +363,7 @@ class BaseDataContainer(PathMixin, AttrsMixin, dantro.abc.AbstractDataContainer)
         """A __format__ helper function: returns info about the items"""
         return "{} stored, {} attributes".format(type(self.data),
                                                  len(self.attrs))
-
-    def convert_to(self, TargetCls, **target_init_kwargs):
-        """With this method, a TargetCls object can be created from this
-        particular container instance.
         
-        Conversion might not be possible if TargetCls requires more information
-        than is available in this container.
-        """
-        log.debug("Converting %s to %s ...", self.logstr, TargetCls.__name__)
-        return TargetCls(name=self.name, attrs=self.attrs, data=self.data,
-                         **target_init_kwargs)
-
 
 # -----------------------------------------------------------------------------
 
@@ -371,13 +373,16 @@ class BaseDataGroup(PathMixin, AttrsMixin, dantro.abc.AbstractDataGroup):
     NOTE: This is still an abstract class and needs to be subclassed.
     """
 
-    def __init__(self, *, name: str, containers: list=None, attrs=None, StorageCls=dict):
-        """Initialize a BaseDataGroup, which can store other containers and attributes.
+    def __init__(self, *, name: str, containers: list=None, attrs=None, StorageCls: type=dict):
+        """Initialize a BaseDataGroup, which can store other containers and
+        attributes.
         
         Args:
             name (str): The name of this data container
-            data (TYPE): The data to store in this container
+            containers (list, optional): The containers to store in this group
             attrs (None, optional): A mapping that is stored as attributes
+            StorageCls (type, optional): in which type of object to store the
+                containers.
         """
         log.debug("BaseDataGroup.__init__ called.")
 
@@ -454,6 +459,7 @@ class BaseDataGroup(PathMixin, AttrsMixin, dantro.abc.AbstractDataGroup):
 
         # Depending on length of the key sequence, start recursion or not
         if len(key) > 1:
+            print(key, self.data)
             self.data[key[0]][key[1:]] = val
             return
         
@@ -478,6 +484,8 @@ class BaseDataGroup(PathMixin, AttrsMixin, dantro.abc.AbstractDataGroup):
         if len(key) > 1:
             # Continue recursion
             del self.data[key[0]][key[1:]]
+            return
+
         # else: end of recursion: delete and unlink this container
         cont = self.data[key[0]]
         del self.data[key[0]]
@@ -599,7 +607,17 @@ class BaseDataGroup(PathMixin, AttrsMixin, dantro.abc.AbstractDataGroup):
 
     def recursive_update(self, other):
         """Recursively updates the contents of this data group with the entries
-        of the given data group"""
+        of the given data group
+        
+        NOTE This will create shallow copies of those elements in `other` that
+        are added to this object.
+        
+        Args:
+            other (BaseDataGroup): The group to update with
+        
+        Raises:
+            TypeError: If `other` was of invalid type
+        """
 
         if not isinstance(other, BaseDataGroup):
             raise TypeError("Can only update {} with objects of classes that "
@@ -618,10 +636,16 @@ class BaseDataGroup(PathMixin, AttrsMixin, dantro.abc.AbstractDataGroup):
                     # Continue recursion
                     self[name].recursive_update(obj)
                 else:
+                    # Can add the object, but need to detach the parent first
+                    # and thus need to work on a copy
+                    obj = copy.copy(obj)
+                    obj.parent = None
                     self.add(obj)
 
             else:
-                # Not a group; add it to this group
+                # Not a group; add a shallow copy
+                obj = copy.copy(obj)
+                obj.parent = None
                 self.add(obj)
 
         log.debug("Finished recursive update of %s.", self.logstr)
@@ -739,14 +763,11 @@ class BaseDataGroup(PathMixin, AttrsMixin, dantro.abc.AbstractDataGroup):
         return self.data.get(key, default)
 
     def setdefault(self, key, default=None):
-        """If `key` is in the dictionary, return its value. If not, insert
-        `key` with a value of `default` and return `default`. `default`
-        defaults to None."""
-        if key in self:
-            return self[key]
-        # else: not available
-        self.data[key] = default
-        return default
+        """This method is not supported for a data group"""
+        raise NotImplementedError("setdefault is not supported by {}! Use the "
+                                  "`add` method or `new_group` and "
+                                  "`new_container` to add elements."
+                                  "".format(self.classname))
 
     # .........................................................................
     # Formatting
@@ -838,13 +859,3 @@ class BaseDataGroup(PathMixin, AttrsMixin, dantro.abc.AbstractDataGroup):
         # Highest level; join the lines together and return that string
         lines.append("")
         return "\n".join(lines)
-
-    # .........................................................................
-    # Conversion
-    
-    def convert_to(self, TargetCls, **target_init_kwargs):
-        """Convert this BaseDataGroup to TargetCls by passing data and attrs"""
-        log.debug("Converting %s '%s' to %s ...", self.classname, self.name,
-                  TargetCls.__name__)
-        return TargetCls(name=self.name, data=self.data, attrs=self.attrs,
-                         **target_init_kwargs)